scratch-notifier: remove persistent script
